Updates test files with new action and reducer names
