+++ src/__tests__/FaceoffsReducer.test.js
@@ -1,0 +1,28 @@
+import { global, faceoffs } from '../state/actions';
+import { reducer } from '../state/reducers/faceoffsReducer';
+
+describe('faceoffsReducer test suite', () => {
+  const initialState = null;
+
+  it('should return the initial state with no actions passed in', () => {
+    const action = { type: null };
+    const state = reducer(initialState, action);
+    expect(state).toEqual(null);
+  });
+
+  it('should return an array if passed in action', () => {
+    const action = {
+      type: faceoffs.SET_SQUAD_FACEOFFS,
+      payload: ['a', 'b', 'c', 'd'],
+    };
+    const state = reducer(initialState, action);
+    expect(state).toHaveLength(4);
+    expect(state).toEqual(['a', 'b', 'c', 'd']);
+  });
+
+  it('should retrun initial state with clear users action', () => {
+    const action = { type: global.CLEAR_USERS };
+    const state = reducer(initialState, action);
+    expect(state).toEqual(null);
+  });
+});
